pyproject.toml
[tool.poetry]
name = "vibecheck"
version = "0.1.0"
description = ""
authors = ["John McDonnell <john.v.mcdonnell@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mastodon-py = "^1.8.1"
sqlalchemy = "^2.0.25"
loguru = "^0.7.2"
instructor = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
